% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envvars.R
\name{shiny_getenv}
\alias{shiny_getenv}
\title{Get Shiny User Variables}
\usage{
shiny_getenv(nms, unset = "", session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{nms}{A character vector of names.}

\item{unset}{The value to return for unset variables.}

\item{session}{A Shiny session object. Defaults to the current session.}
}
\value{
A character vector of values.
}
\description{
Get values from the user data environment of a Shiny session.
}
\examples{
session <- rlang::env(userData = rlang::env(foo = "bar", baz = "foo"))
shiny_getenv(c("foo", "baz"), session = session)
shiny_getenv("qux", unset = "unset", session = session)
}
