% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envvars.R
\name{shiny_setenv}
\alias{shiny_setenv}
\title{Set Shiny User Variables}
\usage{
shiny_setenv(..., session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{...}{Named arguments with values coercible to a string.}

\item{session}{A Shiny session object. Defaults to the current session.}
}
\value{
A logical vector of \code{TRUE} values the same length as \code{...}.
}
\description{
Variables set using \code{\link[=Sys.setenv]{Sys.setenv()}} are available to \emph{all} users of a given
Shiny app. This is not always desirable. \code{\link[=shiny_setenv]{shiny_setenv()}} allows you to set
environment-like variables which are only available to the current user, but
which are available within any Shiny module within the app.
}
\examples{
session <- rlang::env(userData = rlang::env())
shiny_setenv(foo = "bar", baz = "foo", session = session)
session$userData$foo
session$userData[["baz"]]
}
