% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envvars.R
\name{shiny_unsetenv}
\alias{shiny_unsetenv}
\title{Unset Shiny User Variables}
\usage{
shiny_unsetenv(nms, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{nms}{A character vector of names.}

\item{session}{A Shiny session object. Defaults to the current session.}
}
\value{
The modified user data environment, invisibly.
}
\description{
Remove variables from the user data environment of a Shiny session.
}
\examples{
session <- rlang::env(
  userData = rlang::env(foo = "bar", baz = "foo", bar = "baz")
)
shiny_unsetenv(c("foo", "baz"), session = session)
names(session$userData)
}
