% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyenv-package.R
\docType{package}
\name{shinyenv-package}
\alias{shinyenv}
\alias{shinyenv-package}
\title{shinyenv: User-Specific Shiny Environments}
\description{
An experimental package to capture and reroute dangerous Shiny function calls.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shinyworks/shinyenv}
  \item \url{https://shinyworks.github.io/shinyenv/}
  \item Report bugs at \url{https://github.com/shinyworks/shinyenv/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
